Fix thumbnailing files in trash: and recent:
g_file_get_path() will return NULL for files in the trash, and
recent. We're expected to read the G_FILE_ATTRIBUTE_STANDARD_TARGET_URI
attribute to get to the real on-disk file.

We also now check that the path is non-NULL, so as to avoid passing
NULL filenames to libarchive (which seems to make it hang).